backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.5.2
sentence-transformers==2.2.2
torch==2.1.1
transformers==4.35.2
peft==0.6.2
faiss-cpu==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.9
mangum==0.17.0  # For ASGI to Lambda adapter